+++ python/bosdyn-scout/src/bosdyn/scout/client.py
@@ -7,8 +7,7 @@
 """ Scout Client is a client for a single Scout instance. The client uses the Scout web API
     to send HTTPs requests to a number of REStful endpoints using the Requests library.
 """
-from collections.abc import Iterable
-from typing import Dict
+from typing import Dict, Iterable
 
 import requests
 from deprecated.sphinx import deprecated